setting up jest
