+++ .babelrc
@@ -1,0 +1,10 @@
+{
+    "presets": ["@babel/preset-env","@babel/react"],
+    "env":{
+        "test": {
+            "plugins": [
+                "@babel/plugin-proposal-class-properties"
+            ]
+        }
+    }
+}
