+++ package.json
@@ -10,17 +10,44 @@
   },
   "scripts": {
     "start": "webpack-dev-server --config ./webpack.config.js --mode development ./harness/index.js",
-    "build": "webpack --mode production"
+    "build": "webpack --mode production",
+    "test": "jest src --coverage"
+  },
+  "jest": {
+    "verbose": true,
+    "clearMocks": true,
+    "collectCoverage": true,
+    "setupFilesAfterEnv": [
+      "<rootDir>/config/setupTest.js"
+    ],
+    "transform": {
+      "^.+\\.js$": "babel-jest"
+    },
+    "notifyMode": "failure",
+    "coverageThreshold": {
+      "global": {
+        "branches": 50,
+        "functions": 50,
+        "lines": 50,
+        "statements": 50
+      }
+    }
   },
   "devDependencies": {
     "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
+    "babel-jest": "^24.8.0",
     "babel-loader": "^8.0.6",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "css-loader": "^2.1.1",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
+    "jest": "^24.8.0",
     "mini-css-extract-plugin": "^0.7.0",
     "querystringify": "^2.1.1",
     "react-hot-loader": "^4.9.0",
@@ -30,7 +57,6 @@
     "webpack-dev-server": "^3.7.0"
   },
   "dependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
     "intersection-observer": "^0.7.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
