+++ src/components/LazyLoader.js
@@ -1,35 +1,5 @@
 import React from 'react';
-
-class IntersectionObserverSingleton {
-    static allTargets = new Map();
-    static allObservers = new Map();
-
-    static getInstance(onIntersection, options, ref) {
-        const key = Object.values(options).join(',');
-        const observer = IntersectionObserverSingleton.allObservers.get(key);
-
-        IntersectionObserverSingleton.allTargets.set(ref, {onIntersection});
-
-        if (!observer) {
-            const newObserver = new IntersectionObserver(IntersectionObserverSingleton.onChange, options);
-            IntersectionObserverSingleton.allObservers.set(key, newObserver);
-            return newObserver;
-        }
-
-        return observer;
-    }
-
-    static onChange(entries) {
-        entries.forEach((entry) => {
-            const {target} = entry;
-            const instance = IntersectionObserverSingleton.allTargets.get(target);
-            if (instance) {
-                instance.onIntersection(entry);
-            }
-        });
-    }
-}
-
+import IntersectionObserverMg  from './IntersectionObserver';
 
 export default class LazyLoader extends React.Component {
     static defaultProps = {
@@ -67,7 +37,7 @@ export default class LazyLoader extends React.Component {
     createObserver = () => {
         const {root, margin, threshold} = this.props;
         const options = {root, margin, threshold};
-        this.observer = IntersectionObserverSingleton.getInstance(this.onIntersection, options, this.target.current);
+        this.observer = IntersectionObserverMg.getInstance(this.onIntersection, options, this.target.current);
         this.observer.observe(this.target.current);
     }
